Added index.jsp and contactForm.jsp, stylesheets and js for them. Changed deletion logic (now records are not deleted from db but they receive the deletion date)

+++ contactsbook/moduleA/src/main/java/DAO/AttachmentDAO.java
@@ -11,7 +11,9 @@ import java.util.List;
  */
 public interface AttachmentDAO {
     void save(Connection conn, Attachment attachment) throws SQLException;
-    List<Attachment> findByContactId(Connection conn, int contactId) throws SQLException;
+
+    List<Attachment> findByContactId(Connection conn, long contactId) throws SQLException;
+
     void delete(Connection conn, Attachment attachment) throws SQLException;
 
 }
